requirements.txt
streamlit
pypdf
PyPDF2
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu      
openpyxl
matplotlib
numpy
pandas